Updated tests and version resolution ordering metric
